+++ src/types.jl
@@ -113,7 +113,7 @@ end
 @define_override_constructor NotebookInfo
 
 abstract type AbstractCancelFlag end
-function is_cancelled(::AbstractCancelFlag) end
+function is_cancelled end
 
 """
     CancelFlag
@@ -403,6 +403,7 @@ const INFERENCE_UNDEF_STATIC_PARAM_CODE = "inference/undef-static-param" # curre
 const INFERENCE_FIELD_ERROR_CODE = "inference/field-error"
 const INFERENCE_BOUNDS_ERROR_CODE = "inference/bounds-error"
 const INFERENCE_METHOD_ERROR_CODE = "inference/method-error"
+const INFERENCE_NON_BOOLEAN_COND_CODE = "inference/non-boolean-cond"
 const TESTRUNNER_TEST_FAILURE_CODE = "testrunner/test-failure"
 
 const ALL_DIAGNOSTIC_CODES = Set{String}(String[
@@ -427,6 +428,7 @@ const ALL_DIAGNOSTIC_CODES = Set{String}(String[
     INFERENCE_FIELD_ERROR_CODE,
     INFERENCE_BOUNDS_ERROR_CODE,
     INFERENCE_METHOD_ERROR_CODE,
+    INFERENCE_NON_BOOLEAN_COND_CODE,
     TESTRUNNER_TEST_FAILURE_CODE,
 ])