+++ test/analysis/test_Analyzer.jl
@@ -318,4 +318,67 @@ end
     end
 end
 
+@testset "NonBooleanCondErrorReport" begin
+    # no report for boolean condition
+    let result = analyze_call((Bool,)) do x
+            x ? 1 : 2
+        end
+        @test isempty(get_reports(result))
+    end
+    let result = analyze_call((Bool,)) do x
+            if x; 1; else; 2; end
+        end
+        @test isempty(get_reports(result))
+    end
+    let result = analyze_call((Bool,)) do x
+            x && return 1
+        end
+        @test isempty(get_reports(result))
+    end
+
+    # basic non-boolean condition
+    let result = analyze_call((Int,)) do x
+            x ? 1 : 2
+        end
+        reports = get_reports(result)
+        @test length(reports) == 1
+        r = only(reports)
+        @test r isa NonBooleanCondErrorReport && r.union_split == 0
+    end
+    let result = analyze_call((Int,)) do x
+            if x; 1; else; 2; end
+        end
+        reports = get_reports(result)
+        @test length(reports) == 1
+        r = only(reports)
+        @test r isa NonBooleanCondErrorReport && r.union_split == 0
+    end
+    let result = analyze_call((Int,)) do x
+            x && return 1
+        end
+        reports = get_reports(result)
+        @test length(reports) == 1
+        r = only(reports)
+        @test r isa NonBooleanCondErrorReport && r.union_split == 0
+    end
+
+    # union split case: only one branch is non-boolean
+    let result = analyze_call((Union{Bool,Int},)) do x
+            x ?  1 : 2
+        end
+        reports = get_reports(result)
+        @test length(reports) == 1
+        r = only(reports)
+        @test r isa NonBooleanCondErrorReport && r.union_split == 2 && length(r.t) == 1
+    end
+
+    # JuliaLang/julia#61526
+    let result = analyze_call((Vector{String},String,)) do xs, x
+            x in tuple(xs) ? 0 : 1
+        end
+        reports = get_reports(result)
+        @test isempty(reports)
+    end
+end
+
 end # module test_LSAnalyzer